fix: ignore create and update outputs for Rule
Currently, when AWS returns the output of CreateRule or UpdateRule,
it returns a list of Actions, which hold TargetGroupArns.
TargetGroups can also be referenced from a k8s TargetGroup object.
When AWS returns the list of Actions, the generated code creates a new
Action struct and populates it with the values from the sdk output
(which do not include the references).
If the user sets a targetGroupRef, we do not have a way to figure out
the correct Action to assign the returned targetGroupRef.

This proposed change ignores setting Actions returned by AWS. This
should be safe to do so, since the Action AWS returns is expected to be
the same one the user defined in their custom resource, ensuring any
references in the field are not lost.